+++ htdocs/js/pages/Alerts.class.js
@@ -548,6 +548,8 @@ Page.Alerts = class Alerts extends Page.PageUtils {
 		var btn = ['text-box-plus-outline', "Create Ticket"];
 		var html = '';
 		
+		var ticket_template = app.config.new_ticket_template || {};
+		
 		var nice_server = this.getNiceServerText(alert.server);
 		var new_subject = `Alert: ${def.title} on ${nice_server}`;
 		
@@ -598,6 +600,18 @@ Page.Alerts = class Alerts extends Page.PageUtils {
 			})
 		});
 		
+		// due
+		html += this.getFormRow({
+			id: 'd_nt_due_preset',
+			content: this.getFormMenuSingle({
+				id: 'fe_nt_due_preset',
+				options: config.ui.new_ticket_dues,
+				value: ticket_template.due || '',
+				auto_add: true,
+				// 'data-shrinkwrap': 1
+			})
+		});
+		
 		html += '</div>';
 		Dialog.confirm( title, html, btn, function(result) {
 			if (!result) return;
@@ -611,11 +625,11 @@ Page.Alerts = class Alerts extends Page.PageUtils {
 				status: 'open',
 				category: '',
 				assignees: $('#fe_nt_assignees').val(),
-				cc: [],
-				notify: [],
+				cc: ticket_template.cc || [],
+				notify: ticket_template.notify || [],
 				events: [],
 				tags: $('#fe_nt_tags').val(),
-				due: '',
+				due: $('#fe_nt_due_preset').val(),
 				server: alert.server || ''
 			};
 			if (!ticket.subject.length) return app.badField('#fe_nt_subject', "Please enter a subject line for the ticket.");
@@ -641,7 +655,7 @@ Page.Alerts = class Alerts extends Page.PageUtils {
 		}); // Dialog.confirm
 		
 		MultiSelect.init( $('#fe_nt_assignees, #fe_nt_tags') );
-		SingleSelect.init( $('#fe_nt_type') );
+		SingleSelect.init( $('#fe_nt_type, #fe_nt_due_preset') );
 		Dialog.autoResize();
 		
 		$('#fe_nt_subject').focus().get(0).setSelectionRange( new_subject.length, new_subject.length );