+++ htdocs/js/pages/admin/Groups.class.js
@@ -1306,20 +1306,6 @@ Page.Groups = class Groups extends Page.ServerUtils {
 		this.div.find('#d_vg_jobs > .box_content').removeClass('loading').html(html);
 	}
 	
-	doAbortJob(id) {
-		// abort job, clicked from active or queued tables
-		Dialog.confirmDanger( 'Abort Job', "Are you sure you want to abort the job &ldquo;<b>" + id + "</b>&rdquo;?", ['alert-decagram', 'Abort'], function(result) {
-			if (!result) return;
-			app.clearError();
-			Dialog.showProgress( 1.0, "Aborting Job..." );
-			
-			app.api.post( 'app/abort_job', { id: id }, function(resp) {
-				Dialog.hideProgress();
-				app.showMessage('success', config.ui.messages.job_aborted);
-			} ); // api.post
-		} ); // confirm
-	}
-	
 	jobActiveNav(offset) {
 		// user clicked on active job pagination nav
 		this.activeOffset = offset;